separate help table from popover
